+++ back-end/hospital-api/src/main/java/net/pladema/imagenetwork/imagequeue/infrastructure/input/rest/dto/ImageQueueFilteringCriteriaDto.java
@@ -19,7 +19,7 @@ public class ImageQueueFilteringCriteriaDto {
     @NotNull
     private DateDto to;
     private Integer equipmentId;
-    private Short modalityId;
+    private Integer modalityId;
     @NotNull
     private List<EImageMoveStatus> statusList;
     private String name;
